+++ PythonProjects/art_gallery/gallery/views.py
@@ -5,10 +5,5 @@ from django.template import loader
 from django.http import HttpResponse
 
 def gallery(request):
-    template = loader.get_template('index.html')
     random_number = random.randint(1, 1000)
-    context = {
-        'random_number': random_number,
-    }
-    return HttpResponse(template.render())
-
+    return render(request, 'index.html', {'random_number': random_number})